Factored out Node and Tree to folder DataStructures
